Revert "Introduce multi-root workspaces support"
This reverts commit b0d7f06cf9ec6fd6b1dcb61dd0273e37dd546ed5.
